webapp/requirements.txt
Flask==2.3.3
uWSGI==2.0.23
pyyaml==6.0.1